+++ CHANGELOG.md
@@ -31,6 +31,8 @@
 * Sixel images moved or deleted on window resize.
 * Cursor sometimes incorrectly restored on exit from alternate screen.
 * 'Underline' cursor being invisible on underlined text.
+* Restored cursor position in 'normal' screen when window was resized
+  while in 'alt' screen.
 
 
 ### Security